Rename delegations entry file
